Readme File
Some information that would be useful to get introduced into the project is to not immediately start 
coding and that it would probably be best to discuss ideas first and past experiences.
This can help ensure the code does not start off messy and scrambled and will guarantee a clean code.
As an administrator who needs to maintain a software product I believe in being able to correct
any errors and constantly make improvements. These are important because correcting bugs is essential in making the program
work properly and consistently with minimal issues.
Being able to improve it is also important as people like to keep things new and exciting
so that is also something that can be very important.

The code is hosted on GitHub and it is written in Python 3 and the only requirements that would be needed to
play this game is to have python installed. In order to run the program you would just run the code and it will
prompt you to make your first move in the game and it will continue as you play. The program is written using if
and else statements to help guide the player through the game smoothly.